One more bugfix; project complete
